requirements.txt
python-telegram-bot~=21.10
tinytuya~=1.16.1
dotenv~=0.9.9
python-dotenv~=1.0.1